requirements.txt
python-pptx>=0.6.23
beautifulsoup4>=4.12.3
lxml>=5.2.2
requests>=2.31.0
pandas>=2.1.3